* merge r240 to r193(new upstream?) * add LDAP deluser support for OSDN unified account (deleteUser() in LDAPDB.pm, is not tested at all)

+++ slashjp/branches/deluser/Slash/LDAPDB/lib/Slash/LDAPDB.pm
@@ -55,17 +55,19 @@ sub new {
 		  @_ };
     bless $self, $class;
 
+    $DEBUG_LEVEL = $constants->{ldap_debug_level} || $DEBUG_LEVEL;
+
     if (!$self->{_disabled} &&
 	!($self->{_ldap} = Net::LDAP->new($self->{host}, timeout => $self->{timeout}))){
-	__debug(1, "LDAP: can't create DLAP object $@");
-	$self->{_disabled} = "DLAP object create fail: $@";
+	__debug(1, "LDAP: can't create LDAP object $@");
+	$self->{_disabled} = "LDAP object create fail: $@";
     }
     return $self;
 }
 
 sub bind {
     my $self = shift;
-    $self->_check_diabled and return undef;
+    $self->_check_disabled and return undef;
     return 1 if defined $self->{_bind_p};
     $self->{_bind_p} = 1;
     my @args = ();
@@ -86,7 +88,7 @@ sub authUser {
     my $matchname = shift;
     my $pass = shift;
     __debug(8, "LDAP::authUser: start auth for $matchname");
-    $self->_check_diabled and return undef;
+    $self->_check_disabled and return undef;
     my $userinfo = $self->getUser($matchname);
 
     if ($userinfo->{passwd} eq md5_hex($pass)) {
@@ -103,7 +105,7 @@ sub createUser {
     my $val  = shift;
     __debug(8, "LDAP::createUser called for user '$user'");
     __debug(9, "LDAP::createUser called with ". Data::Dumper->Dump([$val], [qw($val)]));
-    $self->_check_diabled and return undef;
+    $self->_check_disabled and return undef;
     $val->{matchname} = $user;
 
     unless ($self->getUser($val->{matchname})) {
@@ -141,7 +143,7 @@ sub deleteUserByUid {
     my $self = shift;
     my $uid  = shift;
     __debug(8, "LDAP::deleteUser called for uid '$uid'");
-    $self->_check_diabled and return undef;
+    $self->_check_disabled and return undef;
     my $userent = $self->getUserByUid($uid);
     $userent && $self->deleteUser($userent->{matchname});
 }
@@ -150,8 +152,22 @@ sub deleteUser {
     my $self = shift;
     my $user = shift;
     __debug(8, "LDAP::deleteUser called for user '$user'");
-    $self->_check_diabled and return undef;
-    my $mesg = $self->_timeout(sub { $self->{_ldap}->delete("cn=${user},".$self->{base_dn}) });
+    $self->_check_disabled and return undef;
+
+    my $userinfo = $self->getUser($user);
+    if (grep('otpUserInfo', $userinfo->{objectClass})) {
+      __debug(8, "LDAP::deleteUser: User $user is also OTP's. The LDAP entry is only modified.");
+      my $mesg = $self->_timeout(sub { $self->{_ldap}->modify("cn=$user,$self->{base_dn}",
+							      changes => [
+									  delete => [
+										     objectClass => 'slashdotUserInfo',
+										     'slashdotUidNumber',
+										     'slashdotEmail'
+										    ]
+									 ]) });
+    } else {
+      my $mesg = $self->_timeout(sub { $self->{_ldap}->delete("cn=${user},".$self->{base_dn}) });
+    }
     $mesg->code && __debug(3, "LDAP Error when deleteUser: ". $mesg->error);
     !$mesg->code;
 }
@@ -162,11 +178,11 @@ sub setUser() {
     my $val  = shift;
     __debug(8, "LDAP::setUser called for user '$user'");
     __debug(9, "LDAP::setUser called with ". Data::Dumper->Dump([$val], [qw($val)]));
-    $self->_check_diabled and return undef;
+    $self->_check_disabled and return undef;
     my $changes = $self->_s2lop($val);
     return -1 unless @$changes;
 
-    __debug(9, "DLAP::serUser ".Data::Dumper->Dump([$changes], [qw($changes)]));
+    __debug(9, "LDAP::setUser ".Data::Dumper->Dump([$changes], [qw($changes)]));
 
     $self->bind;
     my $mesg = $self->_timeout(sub { $self->{_ldap}->modify("cn=$user,$self->{base_dn}",
@@ -181,7 +197,7 @@ sub setUserByUid() {
     my $val  = shift;
     __debug(8, "LDAP::setUserByUid called for uid '$uid'");
     __debug(9, "LDAP::setUserByUid called with ". Data::Dumper->Dump([$val], [qw($val)]));
-    $self->_check_diabled and return undef;
+    $self->_check_disabled and return undef;
     my $changes = $self->_s2lop($val);
     return -1 unless @$changes;
 
@@ -193,7 +209,7 @@ sub setUserByUid() {
 sub getUser() {
     my $self = shift;
     my $user = shift;
-    $self->_check_diabled and return undef;
+    $self->_check_disabled and return undef;
     my $entry = $self->_get_userent("(cn=$user)");
     unless ($entry) {
 	__debug(5, "LDAP::getUser: Can't find user cn=$user");
@@ -205,7 +221,7 @@ sub getUser() {
 sub getUserByUid() {
     my $self = shift;
     my $uid  = shift;
-    $self->_check_diabled and return undef;
+    $self->_check_disabled and return undef;
     my $entry = $self->_get_userent("($self->{attrib_prefix}UidNumber=$uid)");
     unless ($entry) {
 	__debug(5, "LDAP::getUser: Can't find user uid=$uid");
@@ -311,7 +327,7 @@ sub __getCurrentStatic {
     Slash::Utility::Environment::getCurrentStatic();
 }
 
-sub _check_diabled {
+sub _check_disabled {
     my $self = shift;
     $self->{_disabled} and
 	__debug(1, "Slash::LDAP has been disabled: ($self->{_disabled})");
